requirements.txt
langchain==0.0.202
python-dotenv==1.0.0
ffmpeg-python==0.2.0
tiktoken==0.4.0
streamlit==1.23.1